+++ src/features/update-comment/api/deleteComment.ts
@@ -1,0 +1,6 @@
+import api from '@/shared/api/axios';
+
+export const deleteComment = async (commentId: number) => {
+  const response = await api.delete(`/feeds/comments/${commentId}`);
+  return response.data;
+};
