+++ src/features/update-comment/api/postComment.ts
@@ -5,8 +5,3 @@ export const postComment = async (feedId: number, body: CommentReq) => {
   const response = await api.post<{ id: number }>(`/feeds/${feedId}/comments`, body);
   return response.data;
 };
-
-export const deleteComment = async (commentId: number) => {
-  const response = await api.delete(`/feeds/comments/${commentId}`);
-  return response.data;
-};
